feat: created separate library for TimeSlider
