+++ libs/shared/design-system/src/index.ts
@@ -23,4 +23,3 @@ export { Notification } from './lib/notification/notification';
 export { Error } from './lib/error/error';
 export { Tooltip } from './lib/tooltip/tooltip';
 export { useOutsideClick } from './lib/hooks/use-outside-click';
-export { TimeSlider } from './lib/time-slider/time-slider';
