+++ libs/shared/ui/time-slider/README.md
@@ -1,0 +1,7 @@
+# TimeSlider
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test TimeSlider` to execute the unit tests via [Vitest](https://vitest.dev/).
