+++ libs/shared/ui/time-slider/project.json
@@ -1,0 +1,19 @@
+{
+  "name": "TimeSlider",
+  "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/shared/ui/time-slider/src",
+  "projectType": "library",
+  "tags": ["type:shared/ui", "scope:shared"],
+  "targets": {
+    "lint": {
+      "executor": "@nx/eslint:lint"
+    },
+    "test": {
+      "executor": "@nx/vite:test",
+      "outputs": ["{options.reportsDirectory}"],
+      "options": {
+        "reportsDirectory": "../../../../coverage/libs/shared/ui/time-slider"
+      }
+    }
+  }
+}
