+++ libs/shared/ui/time-slider/src/index.ts
@@ -1,0 +1,1 @@
+export { TimeSlider } from './lib/time-slider';
