+++ libs/shared/ui/time-slider/src/lib/time-slider.tsx
