+++ libs/shared/utils/date/src/index.ts
@@ -7,5 +7,6 @@ export {
   dateToNumber,
   type TDateTimeString,
   type TDateString,
+  type TDateStringInternal,
 } from './lib/date.utils';
 export { formatHour, type THourFormat } from './lib/time.utils';
